made server.py modular and put all css in style.css instead of in the htmls
