+++ flag_generators/gen_17_nmap_scanning.py
@@ -11,13 +11,14 @@ from flag_generators.flag_helpers import FlagUtils
 class NmapScanFlagGenerator:
     """
     Generator for the Nmap Scanning challenge.
-    Dynamically patches web_version_admin/server.py with random ports,
+    Dynamically patches web_version_admin/fake_services.py with random ports,
     flags, service names, and junk responses. Stores unlock metadata in memory.
     """
 
-    def __init__(self, project_root: Path = None, server_file: Path = None, mode="guided"):
+    def __init__(self, project_root: Path = None, services_file: Path = None, mode="guided"):
         self.project_root = project_root or self.find_project_root()
-        self.server_file = server_file or self.project_root / "web_version_admin" / "server.py"
+        # CHANGED: Now points to fake_services.py instead of server.py
+        self.services_file = services_file or self.project_root / "web_version_admin" / "fake_services.py"
         self.mode = mode
         self.port_range = list(range(8000, 8100)) if self.mode == "guided" else list(range(9000, 9100))
         self.metadata = {}
@@ -38,18 +39,20 @@ class NmapScanFlagGenerator:
     def escape_python_string_literal(self, s: str) -> str:
         return '"' + s.replace('\\', '\\\\').replace('\n', '\\n').replace('"', '\\"') + '"'
 
-    def patch_server_file(self, real_flag: str, fake_flags: dict, real_port: int, junk_ports: dict):
-        server_file = self.server_file.resolve()
+    def patch_services_file(self, real_flag: str, fake_flags: dict, real_port: int, junk_ports: dict):
+        services_file = self.services_file.resolve()
 
-        if not server_file.exists():
-            print(f"❌ ERROR: {server_file} not found.", file=sys.stderr)
+        if not services_file.exists():
+            print(f"❌ ERROR: {services_file} not found.", file=sys.stderr)
             sys.exit(1)
 
         try:
+            # Determine variable names based on mode
             flag_var = "GUIDED_FAKE_FLAGS" if self.mode == "guided" else "SOLO_FAKE_FLAGS"
             junk_var = "GUIDED_JUNK_RESPONSES" if self.mode == "guided" else "SOLO_JUNK_RESPONSES"
             service_var = "GUIDED_SERVICE_NAMES" if self.mode == "guided" else "SOLO_SERVICE_NAMES"
 
+            # Prepare data
             all_ports = [real_port] + list(fake_flags.keys()) + list(junk_ports.keys())
             service_name_pool = [
                 "alpha-core", "delta-sync", "gamma-relay", "beta-hub", "lambda-api", "omega-stream",
@@ -62,6 +65,7 @@ class NmapScanFlagGenerator:
                 for i, port in enumerate(all_ports)
             }
 
+            # Build source code strings for the new dictionaries
             new_fake_flags = [f"    {real_port}: \"{real_flag}\",       # ✅ REAL FLAG"]
             for port, flag in fake_flags.items():
                 new_fake_flags.append(f"    {port}: \"{flag}\",       # fake")
@@ -76,12 +80,13 @@ class NmapScanFlagGenerator:
             service_name_updates = [
                 f"    {port}: \"{combined_service_names[port]}\"" for port in sorted(combined_service_names)
             ]
-            new_service_names_block = f"{service_var}.update({{\n" + ",\n".join(service_name_updates) + "\n}})"
-            base_service_block = f"{service_var} = {{\n" + ",\n".join(service_name_updates) + "\n}"
+            new_service_names_block = f"{service_var} = {{\n" + ",\n".join(service_name_updates) + "\n}"
 
-            print(f"📂 Reading {server_file}...")
-            content = server_file.read_text(encoding="utf-8")
+            print(f"📂 Reading {services_file}...")
+            content = services_file.read_text(encoding="utf-8")
 
+            # --- Apply Regex Replacements ---
+            # Replace FAKE_FLAGS dictionary
             new_content, count_flags = re.subn(
                 rf"{flag_var}\s*=\s*\{{[^}}]*\}}",
                 new_fake_flags_block,
@@ -89,8 +94,9 @@ class NmapScanFlagGenerator:
                 flags=re.DOTALL
             )
             if count_flags == 0:
-                raise RuntimeError(f"No {flag_var} block found to replace!")
+                raise RuntimeError(f"No {flag_var} block found in {services_file.name} to replace!")
 
+            # Replace JUNK_RESPONSES dictionary
             new_content, count_junk = re.subn(
                 rf"{junk_var}\s*=\s*\{{[^}}]*\}}",
                 new_junk_block,
@@ -98,49 +104,38 @@ class NmapScanFlagGenerator:
                 flags=re.DOTALL
             )
             if count_junk == 0:
-                raise RuntimeError(f"No {junk_var} block found to replace!")
+                raise RuntimeError(f"No {junk_var} block found in {services_file.name} to replace!")
 
-            new_content, count_base_services = re.subn(
+            # Replace SERVICE_NAMES dictionary
+            new_content, count_services = re.subn(
                 rf"{service_var}\s*=\s*\{{[^}}]*\}}",
-                base_service_block,
-                new_content,
-                flags=re.DOTALL
-            )
-            if count_base_services == 0:
-                print(f"⚠️ WARNING: No base assignment for {service_var} found — skipping.")
-
-            new_content, count_update_block = re.subn(
-                rf"{service_var}\.update\(\s*\{{[^}}]*\}}\s*\)",
                 new_service_names_block,
                 new_content,
                 flags=re.DOTALL
             )
-            if count_update_block == 0:
-                new_content = new_content.replace(
-                    "# Combine real flags and junk responses",
-                    "# Combine real flags and junk responses\n" + new_service_names_block
-                )
+            if count_services == 0:
+                print(f"⚠️ WARNING: No {service_var} definition found — skipping service name update.")
 
-            backup_file = server_file.with_suffix(".bak")
+            # Create backup
+            backup_file = services_file.with_suffix(".bak")
             if not backup_file.exists():
-                server_file.replace(backup_file)
+                services_file.replace(backup_file)
                 print(f"🗄️ Backup created: {backup_file.name}")
-            else:
-                print(f"🗄️ Backup already exists: {backup_file.name}")
 
-            server_file.write_text(new_content, encoding="utf-8")
-            print(f"✅ Updated {server_file.name} with {flag_var}, {junk_var}, and {service_var} data.")
+            # Write changes
+            services_file.write_text(new_content, encoding="utf-8")
+            print(f"✅ Updated {services_file.name} with {flag_var}, {junk_var}, and {service_var} data.")
 
             self.metadata = {
                 "real_flag": real_flag,
                 "real_port": real_port,
-                "server_file": str(server_file.relative_to(self.project_root)),
+                "server_file": str(services_file.relative_to(self.project_root)),
                 "unlock_method": f"Scan ports {self.port_range[0]}–{self.port_range[-1]} and query HTTP endpoints to locate the real flag (port {real_port})",
                 "hint": f"Use nmap -p{self.port_range[0]}-{self.port_range[-1]} localhost to discover ports and curl to check flags."
             }
 
         except Exception as e:
-            print(f"❌ ERROR during server.py patching: {e}", file=sys.stderr)
+            print(f"❌ ERROR during fake_services.py patching: {e}", file=sys.stderr)
             sys.exit(1)
 
     def generate_flag(self, challenge_folder: Path) -> str:
@@ -169,13 +164,13 @@ class NmapScanFlagGenerator:
         selected_junk_ports = self.random_ports(self.port_range, selected_flag_ports, random.randint(8, 12))
         junk_responses = {port: random.choice(junk_response_pool) for port in selected_junk_ports}
 
-        self.patch_server_file(real_flag, fake_flags, real_port, junk_responses)
+        self.patch_services_file(real_flag, fake_flags, real_port, junk_responses)
 
         print(f"🏁 Real flag: {real_flag} on port {real_port}")
         for port, flag in fake_flags.items():
             print(f"🎭 Fake flag: {flag} on port {port}")
 
-        # Save unlock metadata just like other generators
+        # Save unlock metadata
         unlocks_file = self.project_root / "web_version_admin" / (
             "validation_unlocks.json" if self.mode == "guided" else "validation_unlocks_solo.json"
         )
